requirements.txt
alembic==1.9.2
aniso8601==8.0.0
bcrypt==4.0.1
certifi==2020.11.8
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cycler==0.11.0
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.37.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.5.3
mccabe==0.6.1
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2020.1
requests==2.29.0
six==1.15.0
SQLAlchemy==2.0.9
toml==0.10.1
typing_extensions==4.5.0
urllib3==1.26.1
Werkzeug==2.2.3
wrapt==1.12.1
WTForms==3.0.1
